exercise 4.21 blog list expansion step 9

+++ part4/bloglist/controllers/blog.js
@@ -4,7 +4,6 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 const JWT_SECRET = require('../utils/secrets').JWT_SECRET
 
-
 blogRouter.get('/', async (request, response) => {
 	const result = await Blog.find({}).populate('user', { name: 1, username: 1 })
 	return response.json(result)
@@ -54,19 +53,40 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+	const id = request.params.id
+
 	try {
-		const id = request.params.id
-		const result = await Blog.findByIdAndDelete(id)
-		if (!result) {
+		const decodedToken = jwt.verify(request.token, JWT_SECRET)
+		if (!decodedToken.id) {
+			return response
+				.status(401)
+				.json({ error: 'Unauthorized', message: 'Invalid token' })
+		}
+
+		const blogToDelete = await Blog.findById(id)
+
+		if (!blogToDelete) {
 			return response.status(204).end()
 		}
-		return response
-			.status(200)
-			.json({ message: `a blog with id ${result._id} has been deleted` })
+
+		if (blogToDelete.user.toString() === decodedToken.id) {
+			const result = await Blog.findByIdAndDelete(id)
+			return response
+				.status(200)
+				.json({ message: `a blog with id ${result._id} has been deleted` })
+		} else {
+			return response.status(403).json({
+				error: 'Forbidden',
+				message: 'You do not have enough permissions',
+			})
+		}
 	} catch (error) {
 		if (error.name === 'CastError') {
 			return response.status(400).json({ error: 'bad id' })
 		}
+		return response
+			.status(401)
+			.json({ error: error.name, message: error.message })
 	}
 })
